Better names and refactoring
